Rework user environment
This is a reaction of the comments in
<https://github.com/flyingcircusio/nixpkgs/pull/15>. It puts ab,
htpasswd, dig and others back into the path.

I think it is also a good idea to enable nixosManual. Without it, all
nixos-* commands don't even react to --help. We should invest slightly
larger build times for an improved user experience.

Re #20505