Get rid of policy routing
Add 'simple routing' scheme which just installs device routes for all
subnets present on a segment -- regardless of the presence of an address
for that specific subnet (filtered by AF though).

Kernel policy routing tables are generally not used with this approach.

Still need to implement a feature switch.

Re #20617