Introduce `fc-manage --build`
fc-manage --build rebuilds either the channel or the development build
depending on the last explicit selection. The latter is determined from
the presence of a channel in ~/.nix-channels.

Add sudo rules so that service users can now invoke `sudo fc-manage
--build` instead of `sudo systemctl restart fc-manage`.

Also relax locking and argument checking to improve flexibility.

Re #23731