Increase kernel loglevel to 4.
At 0 we do not see kernel dumps on the serial console when it's not absolutely
a panic. We've seen VMs that were in a semi-usable state, probably due to
some storage issue, but it's unclear whether Ceph or XFS was at fault.
I'm betting the kernel is saying _something_ but we can't see it.