Reviewed our original Gentoo kernel config and integrated them.
Specifically this will enable the software watchdog in VMs and it also
ensures a couple of other settings that we got by experience.

I wasn't able to port most of the "disable this subsystem" statements as
the NixOS base configuration contains dependent configs to things like ATA,
Wireless, Sound, and others. :/

However, I also decided that using our own base config would be a bad idea.
Disabling those subsystems was helping us to reduce compile times which
isn't directly an issue any longer. However, I like to repeat those settings
that we truly value -- that way we notice if the NixOS baseline does something
that we think may hurt us.