Skip warm reboots if there is a cold boot pending
Imagine a VM sees both a kernel rebuild (triggering a warm reboot) and a
change of Qemu's binary generation (triggering a cold reboot). In this
case, the warm reboot is performed and the cold reboot degenerates into
a no-op since the boottime has changed. But we really want to skip the
warm reboot and perform the cold boot.

Extend reboot code so that warm reboots decline execution if there is
also a cold boot in the queue.

To get this done, we amend Activity and Request objects with
backpointers to their repective Request and ReqManager objects so that
Activities are able to see what else is in the queue.

Re #23965