Improve testing: disable domain and FC agent
I think that the domain "gocept.net" is specific to the FC
infrastructure and should not be set outside, e.g. in tests or Vagrant
boxes.

Disable fc-manage in test VMs since it won't be needed in most cases. It
just adds to the overhead. If you need it, just enable it in your
specific test setup.

Re #23966