FCIO_* should not depend on the existence of ENC data
We assume that FCIO_* mirrors exactly what the directory thinks (or at
least, what is present in the local ENC data copy), and not what is
configured on the local system by chance. So these variables should not
be set in case that ENC data cannot be read.

Also fix keys import script and systemd dependency:

- The import script had several minor defects like missing path names.

- systemd tried to start graylog-update-geolite before graylog. This
  fails when `/var/lib/graylog` has not been created before.

Re #22495