Create a proxy for collectd data which compresses
The idea is to aggregate data per location, and to compress it before sending it over the wire. Compressing saves about 9/10 of bandwidth here.

Currently collectd is configured to send to both stats.flyingcircus.io and a local proxy. This wasy we can see how much resources the compression requires before actually putting it into production. Luckly influxdb depulicates the data for the case that we get it twice.