Unblacklist the i6300esb module to allow the watchdog to work.
We can't prove that this does anything useful, but we bashed our heads for
about 30 minutes trying to provoke a watchdog reboot. Unfortunately, the
kexec mechanism that is in place usually catches all issues we can provoke
faster than the watchdog. In general this is good news because that means
we have multiple safety nets. In practice we can only show that the module
now loads and that systemd is picking it up.

Fixes #23942