Fix Hydra evaluation error
```
in job ‘tested’:
overrideDerivation not yet supported for functors
```

The `nixpkgs` Hydra job tries to build all attributes from the attrset
returned by all-packages.nix. Alas, not all attributes are derivations.
Some are sub-attrsets, some are lambdas (like `buildRustPackage`). We
now filter the attrset so that only derivations remain in the Hydra job.