Accept failure of `sa1 --boot`
It is ok when the invocation of `sa1 --boot` does not return with exit
status 0 - the most common cause is that systemd gets restarted and the
files are still locked from regular running sadc instances.

Should `sa1 --boot` fail when it is _really_ needed, subsequent
invocations of sadc will also fail. Thus this condition will be reported
in monitoring anyway.