Fix activation mode of sysstat-collect.service
sysstat-collect has a long-running non-forking process but was marked as
"oneshot" service. This leads to systemd blocking dependend start/stop
operations until the spawned sadc process terminates, which can take a
few minutes.

During service activation, systemd spawns a systed-askpassword process.
This is just a symptom and not a cause.