Fix edge case in local firewall rule generation
In case there is nothing in `/etc/local/firewall`, the derivation
`localRules` produces an empty output. It seems that Hydra is not able
to serve the corresponding NAR correctly. Marking the derivation as
local build (depends on stuff in /etc/local) and skipping evaluation
completely in case of no local firewall rules.

Re #29642