Fix unconfigured ES case
The logic instroduced in 2a1bdee was flawed in the respect that it
failed to evaluate if no ElasticSearch version was set.

Example Hydra error:

in job ‘nixos.tests.flyingcircus.nodejs_4’:
value is null while a string was expected

in job ‘flyingcircus_vm_image’:
value is null while a string was expected
[...]

I believe this came from the string interpolations in line 18 and 41.

Re #100251