Perform lazy activation: do not re-activate current system versions repeatedly.
This should reduce the amount of stress on our systemd setups that sometimes
seem to interfere with short term IO stalls and result in unexpected reboots.

However, this also removes the automatic re-activation of crashed services.
But we should be properly restarting them anyway and not rely on this
accidental behaviour.